refactor: 질문 페이지의 코드를 컴포넌트화하여 구조 모듈화
- /q/[qid]/page.jsx 파일에서 코드를 추출하고 재사용 가능한 컴포넌트로 모듈화:
  - 게시물 메타데이터를 표시하는 PostInfoPanel.jsx 컴포넌트 생성
  - Quill을 사용하여 질문 내용을 렌더링하는 QuillViewer.jsx 컴포넌트 생성
  - 향후 사용할 수 있는 공통 Modal 컴포넌트 (Modal.jsx) 추가
  - 각 컴포넌트에 맞는 CSS 파일 업데이트
- 참고: 댓글 관련 코드는 여전히 더미 상태이며 추가 정리가 필요함.
