--- src/app/questions/[questionId]/page.module.css
@@ -1,61 +1,0 @@
-.title {
-    margin-left: 12px;
-}
-
-.info {
-    display: flex;
-    justify-content: space-between; /* 왼쪽은 이메일과 메타 정보, 오른쪽은 삭제 버튼 */
-    align-items: center; /* 세로 정렬을 중앙으로 */
-    font-size: 16px;
-    margin: 12px;
-}
-
-.email {
-    font-weight: bold;
-}
-
-.metaInfo {
-    color: #6c757d;
-}
-
-.content {
-    font-size: 18px; /* 원하는 폰트 사이즈로 설정 */
-    line-height: 1.5; /* 줄 간격 조정 (선택 사항) */
-}
-
-.button {
-    margin-left: auto; /* 자동 여백을 사용하여 오른쪽으로 배치 */
-}
-
-.button button:first-child {
-    margin-right: 10px; /* 수정 버튼과 삭제 버튼 사이의 간격 */
-}
-
-/* 숨기기 위한 스타일 */
-.hide {
-    visibility: hidden;
-}
-
-/* styles/page.module.css */
-.modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.modalContent {
-    background-color: white;
-    padding: 20px;
-    border-radius: 5px;
-    text-align: center;
-}
-
-.modalContent button {
-    margin: 10px;
-}
