+++ src/components/question/view/PostInfoPanel.module.css
@@ -1,0 +1,26 @@
+.info {
+    display: flex;
+    justify-content: space-between; /* 왼쪽은 이메일과 메타 정보, 오른쪽은 삭제 버튼 */
+    align-items: center; /* 세로 정렬을 중앙으로 */
+    font-size: 16px;
+    margin: 12px;
+}
+.metaInfo {
+    color: #6c757d;
+}
+.email {
+    font-weight: bold;
+}
+.button {
+    margin-left: auto; /* 자동 여백을 사용하여 오른쪽으로 배치 */
+}
+.button button:first-child {
+    margin-right: 10px; /* 수정 버튼과 삭제 버튼 사이의 간격 */
+}
+/* 숨기기 위한 스타일 */
+.hide {
+    visibility: hidden;
+}
+.title {
+    margin-left: 12px;
+}
