+++ src/components/question/view/QuillViewer.module.css
@@ -1,0 +1,6 @@
+.q-view-content {
+    font-size: 18px;
+    line-height: 1.5;
+    height: 'auto';
+    min-height: 300px;
+}
